+++ packages/leancode_contracts/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+- Fix `TimeOnly` serde
+- Fix `Duration` serde
+
 # 0.1.0
 
 - Initial release